requirements.txt
Django~=4.0
psycopg2~=2.9
djangorestframework~=3.14
htmx